refactor(architecture): extract TransactionItem component & add ErrorBoundary around transaction list
