+++ src/App.css
@@ -140,6 +140,7 @@ body {
 
 .search-icon {
   position: absolute;
+  left: 12px;
   color: #666;
   z-index: 1;
 }
@@ -317,7 +318,8 @@ body {
   background: #f0f0f0;
   padding: 2px 8px;
   border-radius: 4px;
-  display: inline;
+  display: inline-block;
+  width: fit-content;
 }
 
 .transaction-amount {
