+++ src/components/TransactionList.tsx
@@ -5,21 +5,13 @@ import { FixedSizeList as RWFixedSizeList, ListChildComponentProps } from 'react
 const FixedSizeList = RWFixedSizeList as unknown as React.ComponentType<unknown> &
   typeof RWFixedSizeList;
 import { Transaction } from '../types/transaction';
-import { TxType } from '../constants/transactions';
-import { format } from 'date-fns';
+import { TransactionItem } from './TransactionItem';
 
 interface TransactionListProps {
   transactions: Transaction[];
   totalTransactions?: number;
   onTransactionClick: (transaction: Transaction) => void;
 }
-
-const currencyFormatter = new Intl.NumberFormat('en-US', {
-  style: 'currency',
-  currency: 'USD',
-});
-
-const formatCurrency = (amount: number) => currencyFormatter.format(amount);
 
 export const TransactionList: React.FC<TransactionListProps> = ({
   transactions,
@@ -141,102 +133,6 @@ export const TransactionList: React.FC<TransactionListProps> = ({
         >
           {Row}
         </FixedSizeList>
-      </div>
-    </div>
-  );
-};
-
-const TransactionItem: React.FC<{
-  transaction: Transaction;
-  isSelected: boolean;
-  isHovered: boolean;
-  onClick: () => void;
-  onMouseEnter: () => void;
-  onMouseLeave: () => void;
-  rowIndex: number;
-}> = ({ transaction, isSelected, isHovered, onClick, onMouseEnter, onMouseLeave, rowIndex }) => {
-  const formatDate = (date: Date) => {
-    return format(date, 'MMM dd, yyyy HH:mm');
-  };
-
-  const getItemStyle = () => {
-    const accentColor =
-      transaction.type === TxType.Debit ? 'var(--accent-negative)' : 'var(--accent-positive)';
-
-    const base = {
-      backgroundColor: 'var(--card-bg)',
-      borderRadius: '8px',
-      border: '1px solid var(--card-border)',
-      borderLeft: `4px solid ${accentColor}`,
-      padding: '16px',
-      marginBottom: '12px',
-    } as React.CSSProperties;
-
-    if (isHovered || isSelected) {
-      return {
-        ...base,
-        border: `2px solid var(--primary)`,
-        borderLeft: `4px solid var(--primary)`,
-        padding: '15px', // compensate 1px delta each side
-      };
-    }
-
-    return base;
-  };
-
-  return (
-    <div
-      className="transaction-item"
-      style={getItemStyle()}
-      onClick={onClick}
-      onMouseEnter={onMouseEnter}
-      onMouseLeave={onMouseLeave}
-      role="gridcell"
-      aria-rowindex={rowIndex + 1}
-      aria-selected={isSelected}
-      aria-describedby={`transaction-${transaction.id}-details`}
-      tabIndex={0}
-    >
-      <div className="transaction-main">
-        <div className="transaction-merchant">
-          <span className="merchant-name">{transaction.merchantName}</span>
-          <span className="transaction-category">{transaction.category}</span>
-        </div>
-        <div className="transaction-amount">
-          <span className={`amount ${transaction.type}`}>
-            {transaction.type === 'debit' ? '-' : '+'}
-            {formatCurrency(transaction.amount)}
-          </span>
-        </div>
-      </div>
-      <div className="transaction-details" id={`transaction-${transaction.id}-details`}>
-        <div
-          className="transaction-description"
-          aria-label={`Description: ${transaction.description}`}
-        >
-          {transaction.description}
-        </div>
-        <div className="transaction-meta">
-          <span
-            className="transaction-date"
-            aria-label={`Date: ${formatDate(transaction.timestamp)}`}
-          >
-            {formatDate(transaction.timestamp)}
-          </span>
-          <span
-            className={`transaction-status ${transaction.status}`}
-            aria-label={`Status: ${transaction.status}`}
-            aria-live="polite"
-          >
-            {transaction.status}
-          </span>
-          <span
-            className="transaction-location"
-            aria-label={`Location: ${transaction.location ?? 'N/A'}`}
-          >
-            {transaction.location ?? 'N/A'}
-          </span>
-        </div>
       </div>
     </div>
   );
